Move damage back to int from fixed_t, for fixed_T strict cleanup, maybe a couple of real bugs flushed. Only int64_t dodgyness remains.

+++ src/doom/r_plane.cpp
@@ -456,7 +456,7 @@ void R_DrawPlanes (void)
 		    dc_texturemid = dc_texturemid * (textureheight[texture]>>FRACBITS) / SKYSTRETCH_HEIGHT;
 		}
 		flip = (l->special == 272) ? 0 : INT_MAX;
-		an += angle_t(s->textureoffset);
+		an += angle_t((int)s->textureoffset);
 	    }
 	    else
 	    {
